// src/components/group-a/res/view/shared/s-cnt-market-modal-window/res/view/s-cnt-market-specification/res/interface/common.interface.ts
/**
 * Элементы компонента Specification
 */
export interface MarketSpecificationInterface {
  op: string;
  spec: string;
  value: string;
  weight: string;
  protein: string;
  weightProtein: string;
  fat: string;
  weightFat: string;
  carbohydrates: string;
  weightCarbohydrates: string;
  calories: string;
  weightCalories: string;
  item_1: string;
  composition: string;
  info: string;
  brand: string;
  dolce: string;
  manufacturer: string;
  lenta: string;
  country: string;
  russia: string;
  storageConditions: string;
  terms: string;
  term: string;
  month: string;
  kindOfCandy: string;
  chocolate: string;
  typeOfChocolate: string;
  bitter: string;
  flavor: string;
  walnut: string;
  supplements: string;
  hazelnut: string;
  feature: string;
  glazed: string;
  packaging: string;
  box: string;
  hideInformation: string;
}